Addition of ParticularSocActivity
Added ParticularSocActivity and made some changes to EntsActivity to
accommodate the new activity.
ParticularSocActivity  displays a list of all societies. If a user
clicks an item in the list, they are brought to EntsActivity, which then
displays only that society's events.
Navigation drawer has been added to ParticularSocActivity, but is
currently bugged (the presence of the second ListView interferes with
the navigation drawer's click listener).